Excising workflows subdirectory.
