+++ pyproject.toml
@@ -94,13 +94,12 @@ packages = [
     "spatialprofilingtoolbox.workflow.scripts",
     "spatialprofilingtoolbox.workflow.dataset_designs",
     "spatialprofilingtoolbox.workflow.dataset_designs.multiplexed_imaging",
-    "spatialprofilingtoolbox.workflow.workflows",
-    "spatialprofilingtoolbox.workflow.workflows.defaults",
-    "spatialprofilingtoolbox.workflow.workflows.phenotype_proximity",
-    "spatialprofilingtoolbox.workflow.workflows.front_proximity",
-    "spatialprofilingtoolbox.workflow.workflows.density",
-    "spatialprofilingtoolbox.workflow.workflows.nearest_distance",
-    "spatialprofilingtoolbox.workflow.workflows.halo_import",
+    "spatialprofilingtoolbox.workflow.defaults",
+    "spatialprofilingtoolbox.workflow.phenotype_proximity",
+    "spatialprofilingtoolbox.workflow.front_proximity",
+    "spatialprofilingtoolbox.workflow.density",
+    "spatialprofilingtoolbox.workflow.nearest_distance",
+    "spatialprofilingtoolbox.workflow.halo_import",
     "spatialprofilingtoolbox.workflow.common",
     "spatialprofilingtoolbox.workflow.common.logging",
     "spatialprofilingtoolbox.workflow.source_file_adi_parsing",
