+++ spatialprofilingtoolbox/__init__.py
@@ -3,8 +3,8 @@ import pkgutil
 
 from .standalone_utilities.log_formats import colorized_logger
 from .standalone_utilities.configuration_settings import get_version
-from .workflow.workflows import get_workflow
-from .workflow.workflows import get_workflow_names as get_workflow_names  # pylint: disable=useless-import-alias
+from .workflow import get_workflow
+from .workflow import get_workflow_names as get_workflow_names  # pylint: disable=useless-import-alias
 
 
 def get_subpackage_name(module_info):
