+++ spatialprofilingtoolbox/workflow/defaults/computational_design.py
