+++ spatialprofilingtoolbox/workflow/defaults/core.py
