+++ spatialprofilingtoolbox/workflow/defaults/initializer.py
