+++ spatialprofilingtoolbox/workflow/defaults/job_generator.py
