+++ spatialprofilingtoolbox/workflow/front_proximity/core.py
