+++ spatialprofilingtoolbox/workflow/front_proximity/integrator.py
